moves and corrects filepaths
